Add ShadowText class for drawing text to screen
The original game liked to display text with a shadow by drawing the
text twice, once for that shadow and once for the text.

Reimplement that by adding a class for that which wraps some of the
text functions needed and use SFML for drawing text directly getting
rid of the J2ME wrappers for that.